pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "venueflow-ai"
version = "0.1.0"
description = "AI command center for crowd flow, wait times, and live sports venue coordination"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn>=0.30.0",
  "websockets>=14.0",
  "pydantic>=2.8.0",
  "python-dotenv>=1.0.1",
  "openai>=1.40.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
